# metal : refactor mat-vec code