Precursor for dynamic threadgroup allocation.

As a side-effect, some perf gains for `Q5_K` and `IQ4_NL`:

```bash
./scripts/compare-commits.sh master gg/metal-refactor-mv-2 -m ./models/llama-8b-v3-instruct/ggml-model-q5_k.gguf -m ./models/qwen2.5-32b-coder-instruct/ggml-model-q5_k.gguf -m ./models/llama-7b-v2/ggml-model-q5_k_s.gguf -m ./models/llama-7b-v2/ggml-model-q5_k.gguf -m ./models/llama-7b/ggml-model-q5_k_s.gguf -m ./models/llama-7b/ggml-model-q5_k.gguf -m ./models/qwen2.5-3b-coder/ggml-model-q5_k.gguf -m ./models/qwen2.5-7b-coder/ggml-model-q5_k.gguf -m ./models/qwen2.5-32b-instruct/ggml-model-q5_k.gguf -m ./models/qwen2.5-1.5b-coder/ggml-model-iq4_nl.gguf -m ./models/qwen2.5-32b-coder-instruct/ggml-model-iq4_nl.gguf -m ./models/llama-3.1-8b/ggml-model-iq4_nl.gguf -m ./models/qwen2.5-3b-coder/ggml-model-iq4_nl.gguf -m ./models/qwen2.5-7b-coder/ggml-model-iq4_nl.gguf -fa 1 -p 1,1,2 -n 0 -r 10 -t 1
```

| Model                       | Test   |   t/s master |   t/s gg/metal-refactor-mv-2 |   Speedup |
|:----------------------------|:-------|-------------:|-----------------------------:|----------:|
| llama 7B Q5_K_M             | pp1    |        76.58 |                        85.25 |      1.11 |
| llama 7B Q5_K_M             | pp2    |        99.89 |                       114.76 |      1.15 |
| llama 7B Q5_K_S             | pp1    |        76.89 |                        87.37 |      1.14 |
| llama 7B Q5_K_S             | pp2    |        99.16 |                       117.41 |      1.18 |
| llama 8B IQ4_NL - 4.5 bpw   | pp1    |        81.10 |                        93.64 |      1.15 |
| llama 8B IQ4_NL - 4.5 bpw   | pp2    |       111.06 |                       112.73 |      1.02 |
| llama 8B Q5_K_M             | pp1    |        71.98 |                        80.52 |      1.12 |
| llama 8B Q5_K_M             | pp2    |        93.39 |                       106.49 |      1.14 |
| qwen2 1.5B IQ4_NL - 4.5 bpw | pp1    |       168.52 |                       196.51 |      1.17 |
| qwen2 1.5B IQ4_NL - 4.5 bpw | pp2    |       284.07 |                       289.85 |      1.02 |
| qwen2 32B IQ4_NL - 4.5 bpw  | pp1    |        22.36 |                        26.60 |      1.19 |
| qwen2 32B IQ4_NL - 4.5 bpw  | pp2    |        31.62 |                        31.92 |      1.01 |
| qwen2 32B Q5_K_M            | pp1    |        20.27 |                        22.67 |      1.12 |
| qwen2 32B Q5_K_M            | pp2    |        23.87 |                        26.88 |      1.13 |
| qwen2 3B IQ4_NL - 4.5 bpw   | pp1    |       113.93 |                       134.05 |      1.18 |
| qwen2 3B IQ4_NL - 4.5 bpw   | pp2    |       180.21 |                       182.77 |      1.01 |
| qwen2 3B Q5_K_M             | pp1    |       111.91 |                       120.99 |      1.08 |
| qwen2 3B Q5_K_M             | pp2    |       160.14 |                       174.49 |      1.09 |
| qwen2 7B IQ4_NL - 4.5 bpw   | pp1    |        87.34 |                        99.35 |      1.14 |
| qwen2 7B IQ4_NL - 4.5 bpw   | pp2    |       124.08 |                       126.02 |      1.02 |
| qwen2 7B Q5_K_M             | pp1    |        72.05 |                        81.06 |      1.13 |
| qwen2 7B Q5_K_M             | pp2    |        93.17 |                       105.82 |      1.14 |